% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chunk_submit.R
\name{Chunk_submit}
\alias{Chunk_submit}
\title{submit bash commands by user defined number of chunks}
\usage{
Chunk_submit(Commands, Chunk = 6, Name)
}
\arguments{
\item{Commands}{the vector of commands we want to submit (do not need to include nohup portion).}

\item{Chunk}{number of parallel processes.}

\item{Name}{the names of the script and output files.}
}
\examples{

file_names <- grep(".fq$", list.files("/home/zhen/PAR-CLIP/Trim"), value = T)
commands <- paste0("fastq_to_fasta -i ",file_names," -o ", gsub("fq","fa",file_names),"{\\}n",
                  "fastx_collapser -i ",gsub("fq","fa",file_names)," -o ", gsub("trimmed.fq","collapsed.fa",file_names))
Chunk_submit(commands,6,"collapse")

}
