% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Coldata_example}
\alias{Coldata_example}
\title{The example Coldata data.frame}
\format{A \code{data.frame}
\describe{
  As long as the table contains a collumn of the file names, and indicating it is paired or single end library, then it should work fine. 
}}
\usage{
Coldata_example
}
\description{
The example Coldata data.frame
}
\keyword{datasets}
