% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_from_coldata.R
\name{Count_from_coldata}
\alias{Count_from_coldata}
\title{A highly compact but easy to use function to count based on the single based RNA modification GRanges}
\usage{
Count_from_coldata(Coldata, Annot_gr, BAM_dir, title = "test",
  bin_size = 100)
}
\arguments{
\item{Coldata}{a data.frame with 2 collumns neccessarily well defined:

1. a collumn named "SRR_RUN". The entries should be character sting consistant with the name of the bam files (before ".bam").

2. a collumn named "Lib". The entries should be either "Single" or "Paired"}

\item{Annot_gr}{the single based resolution row GRanges used for count.}

\item{BAM_dir}{the directory of your bam files, the function expect all bam files named by "SRR_RUN.bam".}

\item{title}{the title of the file saved.}

\item{bin_zize}{the width of the bin used for count, default is 100.}
}
\value{
The function will output a \code{\link{SummarizedExperiment}} object well defined colData and rowRanges., it will be saved under current working directory.
}
\details{
Under this version, the function is designed to support the count over many (>100) fastq files with both single and paired end data.
If you have only single or paired end data, please use \code{\link{Count_SRRs}}.
}
\examples{

library(RbashGEO)
Annot_GR <- GenomicRanges::reduce( readRDS("hg19_union_gr.rds") ,min.gapwidth=0L)
Coldata_df <- read.csv("ColDesign_Meth.csv")
Count_from_coldata(Coldata_df,Annot_GR,"/home/zhen/bam_Metdb","hg19_union_metdb")

}
\seealso{
\code{\link{Count_SRRs}} is good for count under other contexts.
}
