% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_SRRs.R
\name{Count_SRRs}
\alias{Count_SRRs}
\title{Count data and save the result}
\usage{
Count_SRRs(SRRs, bam_dir = "./", reference_annotation, save_title,
  paired = F, Mode = "Union", Ignore.strand = F, Inter.feature = F)
}
\arguments{
\item{SRRs}{The bam file names (without .bam).}

\item{reference_annotation}{A \code{\link{Granges}} object.}

\item{Mode, Inter.feature, Ignore.strand}{Default settings are "Union",FALSE, and FALSE resectively; see \code{\link{summarizeOverlaps}} for more information.}
}
\examples{

Count_SRRs(SRR_RUN_human_SE,"./",Annotation_gr,F,"Example_human_SE")

}
\seealso{
\code{\link{Rsamtools_view}}
}
