% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsamtools_view.R
\name{Rsamtools_view}
\alias{Rsamtools_view}
\title{Samtools views in R to filter multiple alignments}
\usage{
Rsamtools_view(sam_name, sam_directory = ".", sam_end = ".sam",
  parallel_num = 8, flag_filter = 2816)
}
\examples{
sam_filenames <- gsub(".sam","",grep(".sam$",list.files(),value = T) )
sapply(
 sam_filenames,
 function(x){
   Rsamtools_view(x,flag_filter = 2816) \%>\% Rnohup(.,paste0(x,"_samfilter") )
 }
)
}
