% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_hisat2_reports.R
\name{Check_hisat2_reports}
\alias{Check_hisat2_reports}
\title{Check the hisat2 reports under the curent working directory.}
\usage{
Check_hisat2_reports(result_tail = "_hisat2.out")
}
\examples{
Check_hisat2_reports()

}
