% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhisat2.R
\name{Rhisat2}
\alias{Rhisat2}
\title{Generate hisat2 command in R}
\usage{
Rhisat2(Fastq_file_name = "SRR1234567",
  indx_directory = "/home/zhen/hisat2_idx/hg19/genome",
  Fastq_directory = ".", Paired = F, trim_5 = NULL, trim_3 = NULL,
  parallel_num = 8, Fastq_file_end = ".fastq",
  output_name = Fastq_file_name, tune_Mismatch_MX = NULL,
  tune_Mismatch_MN = NULL)
}
\examples{
#A function to quickly check the alignment efficiencies in R
Hisat2_Output_filenames <- grep("_hisat2.out$",list.files(),value = T)
lst_result <- lapply(Hisat2_Output_filenames,readLines)
SRR_names = gsub("_hisat2.out$","" ,Hisat2_Output_filenames )
names(lst_result) = SRR_names
lst_result

Not_worked <- SRR_names[sapply(lst_result,function(x) length(x) <= 1)]

sapply(
 Not_worked[-1],
 function(x){
   Rhisat2(Fastq_file_name = x,
           Paired = F,
           Fastq_directory = getwd()) \%>\% Rnohup(.,paste0(x,"_hisat2") )
 }


}
