% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgetQC.R
\name{WgetQC}
\alias{WgetQC}
\title{Retrive data from GEO, and run a fastqc}
\usage{
WgetQC(GEO_accession = "SRR1234567", Paired = F)
}
\examples{
sapply(paste0("SRR",3066062:3066069),function(x) nohup_R(WgetQC(x),x))

}
